requirements.txt
notion-client>=1.0.0
notion2md>=3.1.1
requests>=2.28.0
PySide6>=6.4.0
GitPython>=3.1.27
python-dotenv>=0.21.0
keyring>=23.9.3
PyYAML>=6.0
SQLAlchemy>=1.4.42
python-dateutil>=2.8.2
tqdm>=4.64.1
psutil>=5.9.2
pytest>=7.1.3
black>=22.10.0
flake8>=5.0.4
